// src/app/precisa/ventas/Reserva/model/FiltroReserva.ts
export class FiltroReserva {
    constructor() {
    }
    IdReserva:      number;
    IdLote:         number; 
    IdSeparacion:   number; 
    IdManzana:      number;
    IdProyecto:     number;
    IdCliente:      number;
    IdVendedor:     number; 
    CompaniaCodigo: string;
    Codigo:         string;     
    Descripcion:    string;
    IdSede:         number;       
    Cliente:        string;
    Documento:      string; 
    TipoDocumento:  string;
    Estado:         number;
    FechaInicial:    Date;
    FechaFinal:     Date;
}