// src/app/precisa/ventas/Reserva/model/IReserva.ts
export interface IReserva{

    NRO: number;
    DOCUMENTO_CLIENTE :string;
    CLIENTE :string;
    DOCUMENTO_PROMOTOR :string;
    PROMOTOR :string;
    COMPAÑIA: string;
    SUCURSAL: string;
    PROGRAMA: string;
    MANZANA: string;
    LOTE: string;
    FECHA_INICIO: string;
    AREA_M2: string;
    PRECIO: string;
    MONTO_TOTAL: string;
    OBSERVACION: string;
    CUENTABANCARIA: string;
    TIPO_PAGO: string;
    BANCO: string;
    MONEDA: string;
    REFERENCIA: string;
    MONTO_SEPARACION: string;
    FECHA_FIN: string;
    ESTADO_RESERVA: string;
}